+++ engine/src/transposition_table.rs
@@ -5,7 +5,7 @@ use crate::score::Score;
 #[derive(Default, Clone, Debug, PartialEq)]
 pub(crate) struct TranspositionTableEntry {
     pub(crate) best_move: Option<Move>,
-    pub(crate) depth: u16,
+    pub(crate) depth: u8,
     pub(crate) score: Score,
     pub(crate) node_type: NodeType,
 }
